requirements.txt
# Pacvue 智能文档助手 - 依赖库

# LangChain 核心框架
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0

# DeepSeek 模型集成
langchain-deepseek>=0.1.0

# Embeddings 模型
langchain-huggingface>=0.1.0
sentence-transformers>=2.2.0

# 向量存储
chromadb>=0.4.0
langchain-chroma>=0.1.0

# 环境变量管理
python-dotenv>=1.0.0

# Web 服务
fastapi>=0.115.0
uvicorn>=0.32.0
sse-starlette>=2.0.0

# Confluence 集成
requests>=2.31.0
beautifulsoup4>=4.12.0
urllib3>=2.0.0

# MCP (Model Context Protocol) Server
mcp>=1.25.0
